Fix #165
I have added two versions of the same page in the Demo of All Widgets. The first "Employees" is loaded with the full page old which is the equivalent of the application be opened from scratch. The second "Employees (Partial)" does an Ajax load of just the view content. This demonstrates that the flat view control itself is not the performance bottleneck, rather than the initial page load with all of the associated CSS / Javascript files required to make the page work are what impacts performance.